# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: Chatsland Api
# during dev, should point to your local machine
host: localhost:3333
# basePath prefixes all resource paths
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /chats:
    x-swagger-router-controller: Chat
    get:
      tags:
        - 'Chat'
      summary: Get all chats
      operationId: getAllChats
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Chats'
  /chat:
    x-swagger-router-controller: Chat
    post:
      tags:
        - 'Chat'
      summary: Add new chat
      operationId: addNewChat
      parameters:
        - in: 'body'
          name: 'body'
          required: true
          schema:
            $ref: '#/definitions/NewChat'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Chat'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /chat/{id}:
    x-swagger-router-controller: Chat
    get:
      tags:
        - 'Chat'
      summary: Get chat by id
      operationId: getOneChat
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Chat'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - 'Chat'
      summary: Get chat by id
      operationId: updateChat
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - in: 'body'
          name: 'body'
          required: true
          schema:
            $ref: '#/definitions/NewChat'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Chat'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - 'Chat'
      summary: Delete one chat
      operationId: deleteChat
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /channels:
    x-swagger-router-controller: Channel
    get:
      tags:
        - 'Channel'
      summary: Get all channels
      operationId: getAllChannels
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Channels'
  /channel:
    x-swagger-router-controller: Channel
    post:
      tags:
        - 'Channel'
      summary: Add new chat
      operationId: addNewChannel
      parameters:
        - in: 'body'
          name: 'body'
          required: true
          schema:
            $ref: '#/definitions/NewChannel'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Channel'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /channel/{id}:
    x-swagger-router-controller: Channel
    delete:
      tags:
        - 'Channel'
      summary: Delete one channel
      operationId: deleteChannel
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  NewChannel:
    required:
      - chat_id
      - name
      - tags
    properties:
      chat_id:
        type: string
        example: '12sa22as5'
      name:
        type: string
        example: Cafe racers channel
      description:
        type: string
      tags:
        type: array
        items:
          type: string
  Channel:
    properties:
      id:
        type: string
        example: 'asss321'
      chat_id:
        type: string
        example: '12sa22as5'
      name:
        type: string
        example: Cafe racers channel
      description:
        type: string
      tags:
        type: array
        items:
          type: string
  Channels:
    type: array
    items:
      $ref: '#/definitions/Channel'
  NewChat:
    properties:
      name:
        type: string
        example: slack
      icon_url:
        type: string
        example: './img/slack.png'
  Chat:
    properties:
      id:
        type: string
        example: adcfsf125
      name:
        type: string
        example: slack
      icon_url:
        type: string
        example: './img/slack.png'
  Chats:
    type: array
    items:
      $ref: '#/definitions/Chat'
  Message:
    properties:
      message:
        type: string
  ErrorResponse:
    properties:
      message:
        type: string
        example: Error explanation